shift compact filter related code to neutrino node
